requirements1.txt
python-dotenv
requests
langchain-core
langgraph
langchain-huggingface
transformers
torch
typing-extensions
serpapi
